[BISERVER-11724] - Backing out changes made this case.                    Also got the repository test to work
