+++ repository/ivy.xml
@@ -114,7 +114,6 @@
 		<dependency org="junit"                	name="junit"         rev="4.11"     conf="test->default"/>
 		<dependency org="org.slf4j"            	name="slf4j-log4j12" rev="1.6.1"   conf="test->default"/>
 		<dependency org="hsqldb"               	name="hsqldb"        rev="1.8.0.7" conf="test->default"/>
-		<dependency org="org.jmockit"			name="jmockit"		rev="1.9" 		conf="test->default"/>	
 
 
 		<dependency org="org.springframework"  	name="spring-test"   rev="3.2.3.RELEASE"   conf="test->default"/>
